(PC-31098) feat(profile): Fix redirection for ChangeEmail
