+++ src/features/profile/pages/ChangeEmail/ChangeEmail.web.test.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
 
+import { useRoute, replace } from '__mocks__/@react-navigation/native'
 import { UpdateEmailTokenExpiration } from 'api/gen'
 import * as useFeatureFlagAPI from 'libs/firebase/firestore/featureFlags/useFeatureFlag'
 import { mockServer } from 'tests/mswServer'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
-import { act, checkAccessibilityFor, render, screen, waitFor } from 'tests/utils/web'
+import { act, checkAccessibilityFor, render, screen, waitFor, fireEvent } from 'tests/utils/web'
 
 import { ChangeEmail } from './ChangeEmail'
 
@@ -39,6 +40,32 @@ describe('<ChangeEmail/> - old version', () => {
 
         expect(results).toHaveNoViolations()
       })
+    })
+  })
+
+  it('should hide DeleteProfileReasonNewEmailModal when clicking on "Fermer la modale"', async () => {
+    useRoute.mockReturnValueOnce({ params: { showModal: true } })
+    render(reactQueryProviderHOC(<ChangeEmail />))
+
+    const closeButton = screen.getByLabelText('Fermer la modale')
+    fireEvent.click(closeButton)
+
+    await waitFor(() => {
+      expect(screen.queryByText('Modifie ton adresse e-mail sur ce compte')).not.toBeOnTheScreen()
+    })
+  })
+
+  it('should update the URL params when the modal is closed', async () => {
+    useRoute.mockReturnValueOnce({ params: { showModal: true } })
+    render(reactQueryProviderHOC(<ChangeEmail />))
+
+    await screen.findByText('Modifier mon e-mail')
+
+    const closeButton = screen.getByLabelText('Fermer la modale')
+    fireEvent.click(closeButton)
+
+    await waitFor(() => {
+      expect(replace).toHaveBeenNthCalledWith(1, 'ChangeEmail', { showModal: false })
     })
   })
 })
